% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check}
\alias{check}
\title{Check an R package on r-hub}
\usage{
check(path = ".", platform = platforms()$name[1], email = NULL,
  valgrind = FALSE, check_args = character(), show_status = interactive())
}
\description{
Check an R package on r-hub
}

