% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhubv1.R
\name{check_on_linux}
\alias{check_on_linux}
\alias{check_on_windows}
\alias{check_on_macos}
\alias{check_on_debian}
\alias{check_on_ubuntu}
\alias{check_on_fedora}
\alias{check_on_solaris}
\alias{check_on_centos}
\alias{check_with_roldrel}
\alias{check_with_rrelease}
\alias{check_with_rpatched}
\alias{check_with_rdevel}
\alias{check_with_valgrind}
\alias{check_with_sanitizers}
\title{This function is deprecated and defunct. Please see \link{rhubv2}.}
\usage{
check_on_linux(...)

check_on_windows(...)

check_on_macos(...)

check_on_debian(...)

check_on_ubuntu(...)

check_on_fedora(...)

check_on_solaris(...)

check_on_centos(...)

check_with_roldrel(...)

check_with_rrelease(...)

check_with_rpatched(...)

check_with_rdevel(...)

check_with_valgrind(...)

check_with_sanitizers(...)
}
\arguments{
\item{...}{Deprecated.}
}
\description{
This function is deprecated and defunct. Please see \link{rhubv2}.
}
