% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-class.R
\name{get_check}
\alias{get_check}
\title{Retrieve the result of R-hub checks}
\usage{
get_check(ids)
}
\arguments{
\item{ids}{One of the folowing:
\itemize{
\item A single R-hub check id. It may also be the beginning of the hash
of the id (i.e. the part after the last dash), if that check was
listed in the current session.
\item A character vector of check ids. (Or their partial hashes.)
\item An R-hub check group id. (Or its partial hash.)
}}
}
\value{
An \link{rhub_check} object.
}
\description{
Retrieve the result of R-hub checks
}
\seealso{
\code{\link[=list_my_checks]{list_my_checks()}} and \code{\link[=list_package_checks]{list_package_checks()}} to list
R-hub checks.
}
