% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_checks}
\alias{list_checks}
\title{List R-hub builds}
\usage{
list_checks(email = NULL, package = NULL)
}
\arguments{
\item{email}{email address, or \code{NULL}, see details below.}

\item{package}{package name or \code{NULL}, see details below.}
}
\description{
If both \code{email} and \code{package} are \code{NULL}, and the current directory
is an R package, then the checks of this package are queried, using
the maintainer's email address.
}
\details{
If both \code{email} and \code{package} are \code{NULL}, and the current directory
is not an R package, then email address is guessed using
\code{\link[whoami:email_address]{whoami::email_address()}} and all checks that belong to this address
are queried.

If \code{package} is \code{NULL}, but \code{email} is not, then all builds of the
specified email address are queried.

If \code{email} is \code{NULL}, but \code{package} is not, then \code{package} is
interpreted as a path, and the R package at that path is used, with
the maintainer's email address.

If neither \code{email} nor \code{package} are \code{NULL}, then checks for the
specified email address and package are listed.

If you want to use this function programatically, make sure you set
both \code{email} and \code{package}, to avoid context dependent behavior.
}

