% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_my_checks}
\alias{list_my_checks}
\title{List all checks for an email address}
\usage{
list_my_checks(email = email_address(), package = NULL, howmany = 20)
}
\arguments{
\item{email}{Email address. By default it is guessed with
\code{\link[whoami:email_address]{whoami::email_address()}}. The address must be validated, see
\code{\link[=validate_email]{validate_email()}}.}

\item{package}{\code{NULL}, or a character scalar. Can be used to restrict
the search for a single package.}

\item{howmany}{How many checks to show. The current API limit is 20.}
}
\value{
An \code{rhub_check} object.
}
\description{
List all checks for an email address
}
\examples{
\dontrun{
ch <- list_my_checks()
ch
ch$details()
}
}
