% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_package_checks}
\alias{list_package_checks}
\title{List checks of a package}
\usage{
list_package_checks(package = ".", email = NULL)
}
\arguments{
\item{package}{Directory of an R package, or a package tarball.}

\item{email}{Email address that was used for the check(s).
If \code{NULL}, then the maintainer address is used.}
}
\value{
List of check status objects.
}
\description{
List checks of a package
}

