% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{local_check_linux}
\alias{local_check_linux}
\title{Run a package check locally, in a Docker container}
\usage{
local_check_linux(path = ".", quiet = FALSE, image = NULL,
  valgrind = FALSE, check_args = character(), env_vars = character(),
  timeout = Inf, artifacts = tempfile())
}
\description{
Run a package check locally, in a Docker container
}
