% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhubv1.R
\name{platforms}
\alias{platforms}
\title{This function is deprecated and defunct. Please see \link{rhubv2}.}
\usage{
platforms(...)
}
\arguments{
\item{...}{Deprecated.}
}
\description{
This function is deprecated and defunct. Please see \link{rhubv2}.
}
