% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.R
\name{rc_list_local_tokens}
\alias{rc_list_local_tokens}
\title{Show your tokens for the R Consortium runners}
\usage{
rc_list_local_tokens()
}
\value{
Data frame with string columns \code{email} and \code{token}.
}
\description{
Lists all tokens stored on the local machine.
}
\seealso{
Other RC runners API: 
\code{\link{rc_list_repos}()},
\code{\link{rc_new_token}()},
\code{\link{rc_submit}()}
}
\concept{RC runners API}
