% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rc.R
\name{rc_list_repos}
\alias{rc_list_repos}
\title{List your repositories created by the R Consortium runners}
\usage{
rc_list_repos(email = NULL)
}
\arguments{
\item{email}{Email address. We try to detect this, but
if the detection fails, you can specify it explicitly.}
}
\value{
Data frame with columns:
\itemize{
\item \code{repo_name}: Name of the repository.
\item \code{repo_url}: URL of the repository.
\item \code{builds_url}: URL to the builds of the repository.
}

Additional columns and customized printing will be probably added
later to the result.
}
\description{
Lists repositories created by \code{\link[=rc_submit]{rc_submit()}} submissions.
}
\seealso{
Other RC runners API: 
\code{\link{rc_list_local_tokens}()},
\code{\link{rc_new_token}()},
\code{\link{rc_submit}()}
}
\concept{RC runners API}
