% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhub-package.R
\docType{package}
\name{rhub-package}
\alias{rhub}
\alias{rhub-package}
\title{rhub: Connect to 'R-hub'}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Run 'R CMD check' on any of the 'R-hub' architectures, from
    the command line. The current architectures include 'Windows', 'macOS',
    'Solaris' and various 'Linux' distributions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-hub/rhub}
  \item \url{https://r-hub.github.io/rhub/}
  \item Report bugs at \url{https://github.com/r-hub/rhub/issues}
}

}
\author{
\strong{Maintainer}: Gábor Csárdi \email{csardi.gabor@gmail.com}

Authors:
\itemize{
  \item Maëlle Salmon
}

Other contributors:
\itemize{
  \item R Consortium [funder]
}

}
\keyword{internal}
