% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-list-class.R
\name{rhub_check_list}
\alias{rhub_check_list}
\title{An \code{rhub_check_list} object, a list of rhub check results}
\description{
This class is similar to \link{rhub_check} (it inherits from it),
but it shows the results of the checks in a more concise format
by default. To show the details, call the \code{$details()} method.
}
\section{Usage}{
\preformatted{chl$update()
chl$print(...)
chl$details()
chl$livelog(which = 1)
}
}

\section{Details}{


The \code{\link[=list_my_checks]{list_my_checks()}} and \code{\link[=list_package_checks]{list_package_checks()}} functions create
\code{rhub_check_list} objects.

See \link{rhub_check} for the description of the inherited \code{chl$update()},
and \code{chl$livelog()} methods.

\code{chl$details()} prints the details for all checks in the list.

\code{chl$summary()} gives a \code{data.frame} with a line per check indicating \code{package},
\code{version}, \code{status}, \code{submitted}, \code{platform}, \code{group_id} if the check
is part of a group of checks sent simultaneously to the builder (e.g. via
\code{\link[=check_for_cran]{check_for_cran()}}, \code{check_id}.

\code{chl$print()} prints this \code{data.frame}.

\code{chl$get_group(group_id)} returns a \code{\link{rhub_check}} for checks of the
list sharing the \code{group_id}.

\code{chl$get_check(check_id)} returns a \code{\link{rhub_check}} for the check of the
list that has the \code{check_id}.
}

\examples{
\dontrun{
chl <- list_my_checks()
chl
chl$details()
}
}
