% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctor.R
\name{rhub_doctor}
\alias{rhub_doctor}
\title{Check if the current or the specified package is ready to use with R-hub}
\usage{
rhub_doctor(gh_url = NULL)
}
\arguments{
\item{gh_url}{Use \code{NULL} for the package in the current working
directory. Alternatively, use the URL of a GitHub repository that
contains an R package that was set up to use with R-hub.}
}
\description{
Errors if the package or repository is not set up correctly, and
advises on possible solutions.
}
