% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platforms.R
\name{rhub_platforms}
\alias{rhub_platforms}
\title{List R-hub platforms}
\usage{
rhub_platforms()
}
\value{
Data frame with columns:
\itemize{
\item \code{name}: platform name. Use this in the \code{platforms} argument of
\code{\link[=rhub_check]{rhub_check()}}.
\item \code{aliases}: alternative platform names. They can also be used in the
\code{platforms} argument of \code{\link[=rhub_check]{rhub_check()}}.
\item \code{type}:  \code{"os"} or \code{"container"}.
\item \code{os_type}: Linux, macOS or Windows currently.
\item \code{container}: URL of the container image for container platforms.
\item \code{github_os}: name of the OS on GitHub Actions for non-container
platforms.
\item \code{r_version}: R version string. If \code{"*"} then any supported R version
can be selected for this platform.
\item \code{os_name}: name of the operating system, including Linux distribution
name and version for container actions.
}
}
\description{
List R-hub platforms
}
