% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{rhub_setup}
\alias{rhub_setup}
\title{Setup the current R package for use with R-hub}
\usage{
rhub_setup(overwrite = FALSE)
}
\arguments{
\item{overwrite}{if \code{TRUE}, \code{\link[=rhub_setup]{rhub_setup()}} will overwrite an already
existing workflow file.}
}
\value{
Nothing.
}
\description{
It adds or updates the R-hub workflow file to the current package,
and advises on next steps.
}
