% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-cran.R
\name{check_for_cran}
\alias{check_for_cran}
\title{Check an R-package on R-hub, for a CRAN submission}
\usage{
check_for_cran(
  path = ".",
  email = NULL,
  check_args = "--as-cran",
  env_vars = c(`_R_CHECK_CRAN_INCOMING_` = "true", `_R_CHECK_VC_DIRS_` = "true",
    `_R_CHECK_NO_RECOMMENDED_` = "true", `_R_CHECK_REPLACING_IMPORTS_` = "true",
    `_R_CHECK_S3_METHODS_NOT_REGISTERED_` = "true", `_R_CHECK_R_DEPENDS_` = "warn",
    `_R_CLASS_MATRIX_ARRARY_` = "true", `_R_CHECK_FORCE_SUGGESTS_` = "true",
    `_R_CHECK_RD_LINE_WIDTHS_` = "true", `_R_CHECK_EXECUTABLES_EXCLUSIONS_` = "false",
    `_R_CHECK_DOC_SIZES2_` = "true", `_R_CHECK_TIMINGS_` = "10",
    `_R_CHECK_INSTALL_DEPENDS_` = "true", `_R_CHECK_SUGGESTS_ONLY_` = "true",     
    `_R_CHECK_DEPENDS_ONLY_DATA_` = "true", `_R_CHECK_CRAN_INCOMING_` = "true",
    `_R_CHECK_CRAN_INCOMING_REMOTE_` = "true", `_R_CHECK_CODE_ASSIGN_TO_GLOBALENV_` =
    "true", `_R_CHECK_CODE_ATTACH_` = "true", `_R_CHECK_CODE_DATA_INTO_GLOBALENV_` =
    "true", `_R_CHECK_DOT_FIRSTLIB_` = "true", `_R_CHECK_DEPRECATED_DEFUNCT_` = "true",
    `_R_CHECK_SCREEN_DEVICE_` = "stop", `_R_CHECK_WINDOWS_DEVICE_` = "stop",
    `_R_CHECK_TOPLEVEL_FILES_` = "true", `_R_CHECK_LIMIT_CORES_` = "true",
    `_R_CHECK_CODE_USAGE_VIA_NAMESPACES_` = "true",     
    `_R_CHECK_OVERWRITE_REGISTERED_S3_METHODS_` = "true",
    `_R_CHECK_NATIVE_ROUTINE_REGISTRATION_` = "true", `_R_CHECK_NO_STOP_ON_TEST_ERROR_` =
    "true", `_R_CHECK_PRAGMAS_` = "true", `_R_CHECK_COMPILATION_FLAGS_` = "true",
    `_R_CHECK_SERIALIZATION_` = "true", `_R_CHECK_R_ON_PATH_` = "true",
    `_R_CHECK_PACKAGES_USED_IN_TESTS_USE_SUBDIRS_` = "true",
    `_R_CHECK_SHLIB_OPENMP_FLAGS_` = "true", `_R_CHECK_CONNECTIONS_LEFT_OPEN_` = "true",
    `_R_CHECK_FUTURE_FILE_TIMESTAMPS_` = "true", `_R_CHECK_LENGTH_1_CONDITION_` =
    "package:_R_CHECK_PACKAGE_NAME_,[abort,]verbose",      `_R_CHECK_LENGTH_1_LOGIC2_` =
    "package:_R_CHECK_PACKAGE_NAME_,[abort,]verbose", `_R_CHECK_SYSTEM_CLOCK_` = "false",
    `_R_CHECK_AUTOCONF_` = "true", `_R_CHECK_DATALIST_` = "true",
    `_R_CHECK_THINGS_IN_CHECK_DIR_` = "true", `_R_CHECK_THINGS_IN_TEMP_DIR_` = "true",
    `_R_CHECK_BASHISMS_` = "true", `_R_CHECK_ORPHANED_` = "true",
    `_R_CHECK_EXCESSIVE_IMPORTS_` = "20", `_R_CHECK_BOGUS_RETURN_` = "true",
    `_R_CHECK_CRAN_INCOMING_USE_ASPELL_` = "true",
    `_R_CHECK_CRAN_INCOMING_CHECK_FILE_URIS_` = "true"),
  platforms = NULL,
  ...
)
}
\arguments{
\item{path}{Path to a directory containing an R package, or path to
source R package tarball built with \verb{R CMD build} or
\code{devtools::build()}.}

\item{email}{Email address to send notification to about the check.
It must be a validated email address, see \code{\link[=validate_email]{validate_email()}}. If
\code{NULL}, then the email address of the maintainer is used, as defined
in the \code{DESCRIPTION} file of the package.}

\item{check_args}{Arguments for \verb{R CMD check}. By default \code{--as-cran}
is used.}

\item{env_vars}{Environment variables to set on the builder. By default
\verb{_R_CHECK_FORCE_SUGGESTS_=true} is set, to require all packages used.
\verb{_R_CHECK_CRAN_INCOMING_USE_ASPELL_=true} is also set, to use the
spell checker.}

\item{platforms}{A character vector of one or more platforms to build/check
the package on. See \code{\link[=platforms]{platforms()}} for the available platforms. If this is
\code{NULL}, and the R session is interactive, then a menu is shown. If it
is \code{NULL}, and the session is not interactive, then the default R-hub
platforms are used. A vector of platforms which saves time by building one
R package tarball that is used for all the platforms specified.}

\item{...}{Additional arguments are passed to \code{\link[=check]{check()}}.}
}
\value{
An \link{rhub_check} object.
}
\description{
This function calls \code{\link[=check]{check()}} with arguments and platforms, that
are suggested for a CRAN submission.
}
\details{
In particular, if \code{platforms} is \code{NULL} (the default), then
\itemize{
\item It checks the package on Windows, and Linux.
\item It checks the package on R-release and R-devel.
\item It uses the \code{--as-cran} argument to \verb{R CMD check}.
\item It requires all dependencies, including suggested ones.
}

This function is wrapped by \code{devtools::check_rhub()} which you
might find useful if you load \code{devtools} via your .Rprofile (see \code{usethis::use_devtools()}).
}
\examples{
\dontrun{
ch <- check_for_cran("package", show_status = FALSE)
ch$update()
ch$livelog(3)
}
}
