% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-shortcuts.R
\name{check_on_linux}
\alias{check_on_linux}
\alias{check_on_windows}
\alias{check_on_macos}
\alias{check_on_debian}
\alias{check_on_ubuntu}
\alias{check_on_fedora}
\alias{check_on_solaris}
\alias{check_on_centos}
\alias{check_with_roldrel}
\alias{check_with_rrelease}
\alias{check_with_rpatched}
\alias{check_with_rdevel}
\alias{check_with_valgrind}
\alias{check_with_sanitizers}
\title{Check an R package on an R-hub platform}
\usage{
check_on_linux(path = ".", ...)

check_on_windows(path = ".", ...)

check_on_macos(path = ".", ...)

check_on_debian(path = ".", ...)

check_on_ubuntu(path = ".", ...)

check_on_fedora(path = ".", ...)

check_on_solaris(path = ".",
  check_args = "'--no-manual --no-build-vignettes'", ...)

check_on_centos(path = ".", ...)

check_with_roldrel(path = ".", ...)

check_with_rrelease(path = ".", ...)

check_with_rpatched(path = ".", ...)

check_with_rdevel(path = ".", ...)

check_with_valgrind(path = ".", ...)

check_with_sanitizers(path = ".", ...)
}
\arguments{
\item{path}{Path to a directory containing an R package, or path to
source R package tarball built with \code{R CMD build} or
\code{devtools::build()}.}

\item{...}{Additional arguments are passed to \code{\link[=check]{check()}}.}

\item{check_args}{Extra arguments for the \code{R CMD check} command.}
}
\value{
An \link{rhub_check} object.
}
\description{
These functions provide a quick easy to use interface to check a
package on a platform with some particular aspect. Which platform
they use might change over time.
}
