% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhubv1.R
\name{get_check}
\alias{get_check}
\title{This function is deprecated and defunct. Please see \link{rhubv2}.}
\usage{
get_check(...)
}
\arguments{
\item{...}{Deprecated.}
}
\description{
This function is deprecated and defunct. Please see \link{rhubv2}.
}
