% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-class.R
\name{get_check}
\alias{get_check}
\title{Retrieve the result of R-hub checks}
\usage{
get_check(ids)
}
\arguments{
\item{ids}{One of the following:
\itemize{
\item A single R-hub check id.
\item A character vector of check ids.
\item An R-hub check group id.
All ids can be abbreviated, see \link[=rhub-ids]{R-hub ids}.
}}
}
\value{
An \link{rhub_check} object.
}
\description{
Retrieve the result of R-hub checks
}
\section{Examples}{
\preformatted{chk <- get_check("915ee61")
chk
chk$update()
chk$browse()
chk$cran_summary()
chk$urls()
}
}

\seealso{
\code{\link[=list_my_checks]{list_my_checks()}} and \code{\link[=list_package_checks]{list_package_checks()}} to list
R-hub checks.
}
