% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{list_my_checks}
\alias{list_my_checks}
\title{List all checks for an email address}
\usage{
list_my_checks(email = email_address(), package = NULL, howmany = 20)
}
\arguments{
\item{email}{Email address. By default it is guessed with
\code{\link[whoami:email_address]{whoami::email_address()}}. The address must be validated, see
\code{\link[=validate_email]{validate_email()}}.}

\item{package}{\code{NULL}, or a character scalar. Can be used to restrict
the search for a single package.}

\item{howmany}{How many check groups (checks submitted simultaneously)
to show. The current API limit is 20.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} with columns:
\itemize{
\item package Name of the package.
\item version Package version.
\item result: More detailed result of the check. Can be \code{NULL} for errors.
This is a list column with members: \code{status}, \code{errors}, \code{warnings},
\code{notes}.
\item group: R-hub check group id.
\item id: `R-hub check id.
\item platform_name: Name of the check platform.
\item build_time: Build time, a \link{difftime} object.
\item submitted: Time of submission.
\item started: Time of the check start.
\item platform: Detailed platform data, a list column.
\item builder: Name of the builder machine.
\item status Status of the check. Possible values:
\itemize{
\item \code{created}: check job was created, but not running yet.
\item \code{in-progress}: check job is running.
\item \code{parseerror}: internal R-hub error parsing the check results.
\item \code{preperror}: check error, before the package check has started.
\item \code{aborted}: aborted by admin or user.
\item \code{error}: failed check. (Possibly warnings and notes as well.)
\item \code{warning}: \code{R CMD check} reported warnings. (Possibly notes as well.)
\item \code{note}: \code{R CMD check} reported notes.
\item \code{ok}: successful check.
}
\item email: Email address of maintainer / submitter.
}
}
\description{
List all checks for an email address
}
\examples{
\dontrun{
ch <- list_my_checks()
ch
ch$details()
}
}
\seealso{
list_package_checks
}
