% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhubv1.R
\name{list_package_checks}
\alias{list_package_checks}
\title{This function is deprecated and defunct. Please see \link{rhubv2}.}
\usage{
list_package_checks(...)
}
\arguments{
\item{...}{Deprecated.}
}
\description{
This function is deprecated and defunct. Please see \link{rhubv2}.
}
