% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{list_validated_emails}
\alias{list_validated_emails}
\title{List validated email addresses}
\usage{
list_validated_emails()
}
\value{
A data frame with two columns: \code{email} and \code{token}.
If in interactive mode, and there are no validated email addresses,
then a message is printed and the data frame is returned invisibly.
}
\description{
List validated email addresses
}
\seealso{
Other email validation: \code{\link{validate_email}}
}
