% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.R
\name{local_check_linux_images}
\alias{local_check_linux_images}
\title{List R-hub Docker images}
\usage{
local_check_linux_images()
}
\description{
The images are pretty-printed in a short format. Use
\code{as.data.frame()} to get all available platform metadata.
}
