% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{platforms}
\alias{platforms}
\title{List all r-hub platforms}
\usage{
platforms()
}
\description{
The platforms are pretty-printed in a short format. Use
\code{as.data.frame(platforms())} to get all available platform metadata.
}
\examples{
\dontrun{
platforms()
as.data.frame(platforms())
}
}
