% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{status}
\alias{status}
\title{Query the status of an r-hub check}
\usage{
status(id = NULL)
}
\arguments{
\item{id}{The check id, an r-hub status URL, or the object retured
by \code{\link[=check]{check()}}.}
}
\value{
A list with the status of the check. It has entries:
\code{status}, \code{submitted} and \code{duration}. Currently the duration is
only filled when the build has finished.
}
\description{
Query the status of an r-hub check
}

