% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhubv1.R
\name{validate_email}
\alias{validate_email}
\title{This function is deprecated and defunct. Please see \link{rhubv2}.}
\usage{
validate_email(...)
}
\arguments{
\item{...}{Deprecated.}
}
\description{
This function is deprecated and defunct. Please see \link{rhubv2}.
}
