% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/email.R
\name{validate_email}
\alias{validate_email}
\title{Validate an email address on R-hub}
\usage{
validate_email(email = NULL, token = NULL)
}
\arguments{
\item{email}{The email address to validate.}

\item{token}{Token obtained from \code{rhub}, to validate the email address.}
}
\description{
To build and check R packages on R-hub, you need to validate your
email address. This is because R-hub sends out emails about check
results.
}
\details{
The \code{rhub} package stores validated email addresses in a user
configuration file, at a platform-dependent location.
On your current platform the file is at
\Sexpr[stage=render]{rhub:::email_file()}.

To validate a new email address, call this function from an interactive
R session, without any arguments.

To add an email address that was validated before (probably on another
machine), to the configuration file, call this function with the \code{email}
and \code{token} arguments.
}
\seealso{
Other email validation: 
\code{\link{list_validated_emails}()}
}
\concept{email validation}
